+++ database/migrations/2014_10_13_000000_create_invoices_table.php
@@ -21,7 +21,7 @@ class CreateInvoicesTable extends Migration
             $table->string('chauffage')->nullable();
             $table->string('modalite')->nullable();
             $table->date('date')->nullable();
-            $table->tinyInteger('menuiserie')->nullable();
+            $table->integer('menuiserie')->nullable();
 
             $table->double('tva', 8, 2)->default(0)->nullable();
             $table->double('remise', 8, 2)->default(0)->nullable();
